backend/src/config/db.py
from sqlalchemy.ext.asyncio import create_async_engine
from sqlmodel import create_engine
import os

sqlitePath = f"{os.path.abspath('')}/src/db.sqlite3"
sync_engine = create_engine('sqlite:///' + sqlitePath)
async_engine = create_async_engine(
    'sqlite+aiosqlite:///' + sqlitePath, echo=True)
