backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = ""
authors = ["piyushere <kingepiyush@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlmodel = "^0.0.19"
httpx = "^0.26.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.3.1"
aiosqlite = "^0.20.0"
python3-saml = "^1.16.0"
pyjwt = "^2.8.0"
email-validator = "^2.1.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
